requirements/dev.txt
-r base.txt
pytest==3.7.4
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pytest-cache==1.0
pytest-flake8==1.0.2
pytest-mccabe==0.1
pytest-django==3.4.2
